Disallow implicit int conversion

+++ code/qcommon/huffman.c
@@ -390,7 +390,7 @@ void Huff_Decompress(msg_t* mbuf, int offset)
             }
         }
 
-        seq[j] = ch; /* Write symbol */
+        seq[j] = (uint8_t)ch; /* Write symbol */
 
         Huff_addRef(&huff, (uint8_t)ch); /* Increment node */
     }
@@ -420,7 +420,7 @@ void Huff_Compress(msg_t* mbuf, int offset)
     huff.lhead->next = huff.lhead->prev = NULL;
     huff.tree->parent = huff.tree->left = huff.tree->right = NULL;
 
-    seq[0] = (size >> 8);
+    seq[0] = (uint8_t)(size >> 8);
     seq[1] = size & 0xff;
 
     bloc = 16;
